refactor(ui): use LabeledValue for author-labeled metadata blocks
