--- services/dashboard-ui/client/components/deploys/HelmRenderedValues.stories.tsx
@@ -1,57 +1,0 @@
-export default {
-  title: 'Deploys/HelmRenderedValues',
-}
-
-import { HelmDeploymentViewer } from './HelmRenderedValues'
-
-const mockData = {
-  deployments: {
-    default: {
-      'my-app': {
-        metadata: {
-          creationTimestamp: '2024-01-15T10:30:00Z',
-          generation: 3,
-          resourceVersion: '12345',
-          uid: 'dep-abc-123',
-          annotations: {
-            'meta.helm.sh/release-name': 'my-app',
-            'meta.helm.sh/release-namespace': 'default',
-          },
-        },
-        status: {
-          replicas: 3,
-          readyReplicas: 3,
-          availableReplicas: 3,
-          updatedReplicas: 3,
-          conditions: [
-            { type: 'Available', status: 'True', reason: 'MinimumReplicasAvailable' },
-          ],
-        },
-      },
-    },
-  },
-  services: {
-    default: {
-      'my-app': {
-        metadata: {
-          creationTimestamp: '2024-01-15T10:30:00Z',
-          uid: 'svc-abc-123',
-          resourceVersion: '11111',
-        },
-        spec: { type: 'ClusterIP', clusterIP: '10.100.0.1', sessionAffinity: 'None' },
-      },
-    },
-  },
-  ingresses: {},
-  resources: {
-    'my-app-cm': { Kind: 'ConfigMap', Content: 'apiVersion: v1\nkind: ConfigMap\nmetadata:\n  name: my-app-cm' },
-  },
-  manifest: `---
-# Source: my-app/templates/deployment.yaml
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: my-app`,
-}
-
-export const Default = () => <HelmDeploymentViewer data={mockData} />
