+++ services/dashboard-ui/client/components/installs/InstallSyncDetail/index.ts
@@ -1,0 +1,2 @@
+export { InstallSyncDetail } from './InstallSyncDetail'
+export type { IInstallSyncDetail } from './InstallSyncDetail'
