+++ services/dashboard-ui/client/components/workflows/step-details/StepMetadata.tsx
@@ -3,6 +3,7 @@
 import { Code } from '@/components/common/Code'
 import { CodeBlock } from '@/components/common/CodeBlock'
 import { Expand } from '@/components/common/Expand'
+import { LabeledValue } from '@/components/common/LabeledValue'
 import { Status } from '@/components/common/Status'
 import { Text } from '@/components/common/Text'
 import { Time } from '@/components/common/Time'
@@ -89,21 +90,21 @@ export const StepMetadata = ({ step }: IStepDetails) => {
       </Expand>
 
       {step?.status?.metadata?.retry_type && (
-        <div className="flex flex-col gap-1 border rounded-md p-4">
-          <Text variant="label" theme="neutral">
-            Retry info
-          </Text>
-          <Text variant="subtext">
-            Type: {step.status.metadata.retry_type as string}
-          </Text>
-          {step.status.metadata?.retry_idx !== undefined && (
-            <Text variant="subtext">
-              Attempt: {step.status.metadata.retry_idx as number}
-              {step.status.metadata?.max_retries !== undefined
-                ? ` / ${step.status.metadata.max_retries}`
-                : ''}
-            </Text>
-          )}
+        <div className="flex flex-col gap-3 border rounded-md p-4">
+          <Text weight="strong">Retry info</Text>
+          <div className="flex flex-wrap items-start gap-x-16 gap-y-4">
+            <LabeledValue label="Type">
+              {step.status.metadata.retry_type as string}
+            </LabeledValue>
+            {step.status.metadata?.retry_idx !== undefined && (
+              <LabeledValue label="Attempt">
+                {step.status.metadata.retry_idx as number}
+                {step.status.metadata?.max_retries !== undefined
+                  ? ` / ${step.status.metadata.max_retries}`
+                  : ''}
+              </LabeledValue>
+            )}
+          </div>
         </div>
       )}
